pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Amolit <MikeKukPer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
bleach = "^6.2.0"
psycopg2-binary = "^2.9.10"
celery = "^5.4.0"
redis = "^5.2.0"
environs = "^11.2.1"
channels = "^4.2.0"
channels-redis = "^4.2.1"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
websockets = "^14.1"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
